- refactor API request with retry - remove FAST_MODE
