+++ .env.example
@@ -2,7 +2,6 @@ CLOUDFLARE_API_TOKEN=
 CLOUDFLARE_ACCOUNT_ID=
 CLOUDFLARE_LIST_ITEM_LIMIT=300000
 DRY_RUN=0
-FAST_MODE=0
 BLOCK_PAGE_ENABLED=0
 
 # Experimental SNI based filtering. Set to 1 to enable.
