+++ cf_gateway_rule_delete.js
@@ -1,5 +1,5 @@
 import { deleteZeroTrustRule, getZeroTrustRules } from "./lib/api.js";
-import { notifyWebhook } from "./lib/helpers.js";
+import { notifyWebhook } from "./lib/utils.js";
 
 const { result: rules } = await getZeroTrustRules();
 const cgpsRules = rules.filter(({ name }) => name.startsWith("CGPS Filter Lists"));
