+++ cf_list_delete.js
@@ -1,10 +1,8 @@
 import {
-  deleteZeroTrustListsAtOnce,
   deleteZeroTrustListsOneByOne,
   getZeroTrustLists,
 } from "./lib/api.js";
-import { FAST_MODE } from "./lib/constants.js";
-import { notifyWebhook } from "./lib/helpers.js";
+import { notifyWebhook } from "./lib/utils.js";
 
 (async () => {
   const { result: lists } = await getZeroTrustLists();
@@ -31,14 +29,6 @@ import { notifyWebhook } from "./lib/helpers.js";
 
   console.log(`Deleting ${cgpsLists.length} lists...`);
 
-  if (FAST_MODE) {
-    await deleteZeroTrustListsAtOnce(cgpsLists);
-    // TODO: make this less repetitive
-    await notifyWebhook(`CF List Delete script finished running (${cgpsLists.length} lists)`);
-    return;
-  }
-
   await deleteZeroTrustListsOneByOne(cgpsLists);
-
   await notifyWebhook(`CF List Delete script finished running (${cgpsLists.length} lists)`);
 })();
