+++ lib/api.js
@@ -1,5 +1,6 @@
-import { BLOCK_PAGE_ENABLED, LIST_ITEM_SIZE } from "./constants.js";
+import { BLOCK_PAGE_ENABLED, CLOUDFLARE_COOLDOWN_TIME, CLOUDFLARE_REQUEST_RATE_LIMIT, LIST_ITEM_SIZE } from "./constants.js";
 import { requestGateway } from "./helpers.js";
+import { wait } from "./utils.js";
 
 /**
  * Gets Zero Trust lists.
@@ -36,7 +37,9 @@ const createZeroTrustList = (name, items) =>
  * @param {string[]} items The domains.
  */
 export const createZeroTrustListsOneByOne = async (items) => {
+  let requestCount = 1;
   let totalListNumber = Math.ceil(items.length / LIST_ITEM_SIZE);
+  const minutes = CLOUDFLARE_COOLDOWN_TIME / 1000 / 60;
 
   for (let i = 0, listNumber = 1; i < items.length; i += LIST_ITEM_SIZE) {
     const chunk = items
@@ -48,38 +51,22 @@ export const createZeroTrustListsOneByOne = async (items) => {
       await createZeroTrustList(listName, chunk);
       totalListNumber--;
       listNumber++;
+      requestCount++;
+
       console.log(`Created "${listName}" list - ${totalListNumber} left`);
+
+      if (requestCount % CLOUDFLARE_REQUEST_RATE_LIMIT === 0) {
+        console.log(`Waiting for ${minutes} minutes to avoid rate limiting`);
+        await wait(CLOUDFLARE_COOLDOWN_TIME);
+      }
     } catch (err) {
       console.error(`Could not create "${listName}" - ${err.toString()}`);
       throw err;
     }
   }
-};
 
-/**
- * Creates all Zero Trust lists at once.
- * @param {string[]} items The domains.
- */
-export const createZeroTrustListsAtOnce = async (items) => {
-  const requests = [];
-
-  for (let i = 0, listNumber = 1; i < items.length; i += LIST_ITEM_SIZE) {
-    const chunk = items
-      .slice(i, i + LIST_ITEM_SIZE)
-      .map((item) => ({ value: item }));
-    const listName = `CGPS List - Chunk ${listNumber}`;
-
-    requests.push(createZeroTrustList(listName, chunk));
-    listNumber++;
-  }
-
-  try {
-    await Promise.all(requests);
-    console.log("Created lists successfully");
-  } catch (err) {
-    console.error(`Error occurred while creating lists - ${err.toString()}`);
-    throw err;
-  }
+  console.log(`Waiting for ${minutes} minutes to avoid rate limiting`);
+  await wait(CLOUDFLARE_COOLDOWN_TIME);
 };
 
 /**
@@ -99,36 +86,30 @@ const deleteZeroTrustList = (id) =>
  * @param {string} lists[].name The name of a list.
  */
 export const deleteZeroTrustListsOneByOne = async (lists) => {
+  let requestCount = 1;
   let totalListNumber = lists.length;
+  const minutes = CLOUDFLARE_COOLDOWN_TIME / 1000 / 60;
 
   for (const { id, name } of lists) {
     try {
       await deleteZeroTrustList(id);
       totalListNumber--;
+      requestCount++;
+
       console.log(`Deleted ${name} list - ${totalListNumber} left`);
+
+      if (requestCount % CLOUDFLARE_REQUEST_RATE_LIMIT === 0) {
+        console.log(`Waiting for ${minutes} minutes to avoid rate limiting`);
+        await wait(CLOUDFLARE_COOLDOWN_TIME);
+      }
     } catch (err) {
       console.error(`Could not delete ${name} - ${err.toString()}`);
       throw err;
     }
   }
-};
 
-/**
- * Deletes all Zero Trust lists at once.
- * @param {Object[]} lists The lists to be deleted.
- * @param {number} lists[].id The ID of a list.
- * @param {string} lists[].name The name of a list.
- */
-export const deleteZeroTrustListsAtOnce = async (lists) => {
-  const requests = lists.map(({ id }) => deleteZeroTrustList(id));
-
-  try {
-    await Promise.all(requests);
-    console.log("Deleted lists successfully");
-  } catch (err) {
-    console.error(`Error occurred while deleting lists - ${err.toString()}`);
-    throw err;
-  }
+  console.log(`Waiting for ${minutes} minutes to avoid rate limiting`);
+  await wait(CLOUDFLARE_COOLDOWN_TIME);
 };
 
 /**
