+++ lib/constants.js
@@ -26,13 +26,14 @@ export const API_HOST = "https://api.cloudflare.com/client/v4";
 
 export const DRY_RUN = !!parseInt(process.env.DRY_RUN, 10);
 
-export const FAST_MODE = !!parseInt(process.env.FAST_MODE, 10);
-
 export const BLOCK_PAGE_ENABLED = !!parseInt(process.env.BLOCK_PAGE_ENABLED, 10);
 
 export const BLOCK_BASED_ON_SNI = !!parseInt(process.env.BLOCK_BASED_ON_SNI, 10);
 
 export const DEBUG = !!parseInt(process.env.DEBUG, 10);
+
+export const CLOUDFLARE_REQUEST_RATE_LIMIT = 60;
+export const CLOUDFLARE_COOLDOWN_TIME = 2 * 60 * 1000;
 
 export const PROCESSING_FILENAME = {
   ALLOWLIST: "allowlist.txt",
@@ -63,7 +64,7 @@ export const RECOMMENDED_ALLOWLIST_URLS = [
   // Torrent trackers
   "https://raw.githubusercontent.com/im-sm/Pi-hole-Torrent-Blocklist/main/all-torrent-trackres.txt",
   // Banks
-  "https://raw.githubusercontent.com/AdguardTeam/HttpsExclusions/master/exclusions/banks.txt", 
+  "https://raw.githubusercontent.com/AdguardTeam/HttpsExclusions/master/exclusions/banks.txt",
   // Official Discord domains
   "https://raw.githubusercontent.com/Dogino/Discord-Phishing-URLs/main/official-domains.txt",
   // macOS specific
