+++ lib/helpers.js
@@ -5,6 +5,7 @@ import {
   API_KEY,
   API_TOKEN,
 } from "./constants.js";
+import { fetchRetry } from "./utils.js";
 
 if (!globalThis.fetch) {
   console.warn(
@@ -17,58 +18,6 @@ if (!globalThis.fetch) {
     );
   // Import node-fetch since there's no native fetch in this environment
   globalThis.fetch = (await import("node-fetch")).default;
-}
-
-/**
- * Sends a message to a Discord-compatible webhook.
- * @param {url|string} url The webhook URL.
- * @param {string} message The message to be sent.
- * @returns {Promise}
- */
-async function sendMessageToWebhook(url, message) {
-  // Create the payload object with the message
-  // The message is provided as 2 different properties to improve compatibility with webhook servers outside Discord
-  const payload = { content: message, body: message };
-
-  // Send a POST request to the webhook url with the payload as JSON
-  try {
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(payload),
-    });
-
-    // Check if the request was successful
-    if (!response.ok) {
-      throw new Error(`HTTP error! Status: ${response.status}`);
-    } else {
-      return true;
-    }
-  } catch (error) {
-    console.error('Error sending message to webhook:', error);
-    return false;
-  }
-}
-
-/**
- * Sends a CGPS notification to a Discord-compatible webhook.
- * Automatically checks if the webhook URL exists.
- * @param {string} msg The message to be sent.
- * @returns {Promise}
- */
-export async function notifyWebhook(msg) {
-  // Check if the webhook URL exists
-  const webhook_url = process.env.DISCORD_WEBHOOK_URL;
-
-  if (webhook_url && webhook_url.startsWith('http')) {
-    // Send the message to the webhook
-    try {
-      await sendMessageToWebhook(webhook_url, `CGPS: ${msg}`);
-    } catch (e) {
-      console.error('Error sending message to Discord webhook:', e);
-    }
-  }
-  // Not logging the lack of a webhook URL since it's not a feature everyone would use
 }
 
 /**
@@ -87,51 +36,27 @@ const request = async (url, options) => {
   const headers = API_TOKEN
     ? {
         Authorization: `Bearer ${API_TOKEN}`,
-        "Content-Type": "application/json",
       }
     : {
         Authorization: `Bearer ${API_KEY}`,
-        "Content-Type": "application/json",
         "X-Auth-Email": ACCOUNT_EMAIL,
         "X-Auth-Key": API_KEY,
       };
-
-  /*
-	Due to the undocumented Cloudflare API rate limit, the request will be retried until Cloudflare accepts it.
-  */
-  const wait = (ms) => new Promise(resolve => setTimeout(resolve, ms));
-
-  let response;
   let data;
-  let attempts = 0;
-  let maxAttempts = 50;
 
-  while(attempts < 50) {
-    try {
-      response = await fetch(url, {
-        ...options,
-        headers: {
-          ...options.headers,
-          ...headers,
-        },
-      });
+  try {
+    const response = await fetchRetry(url, {
+      ...options,
+      headers: {
+      "Content-Type": "application/json",
+        ...options.headers,
+        ...headers,
+      },
+    });
 
-      data = await response.json();
-
-      if (!response.ok) {
-        throw new Error('Response not OK');
-      }
-      return data;
-    } catch (error) {
-      attempts++;
-      console.warn(`An error occured while making a web request: "${error}", retrying. Attempt ${attempts} of ${maxAttempts}.\nTHIS IS NORMAL IN MOST CIRCUMSTANCES. Refrain from reporting this as a bug unless the script doesn't automatically recover after several attempts.`);
-      if(attempts >= maxAttempts) {
-        // Send a message to the Discord webhook if it exists
-        await notifyWebhook(`An HTTP error has occurred (${response ? response.status : "unknown status"}) while making a web request. Please check the logs for further details.`);
-        throw new Error(`HTTP error! Status: ${response.status} - ${ (data && 'errors' in data) ? data.errors[0].message : data } - ${error}`);
-      }
-      await wait(8000);
-    }
+    return await response.json();
+  } catch (error) {
+    throw new Error(`${(data && 'errors' in data) ? data.errors[0].message : data} - ${error}`);
   }
 };
