+++ lib/utils.js
@@ -66,7 +66,7 @@ export const downloadFiles = async (filePath, urls) => {
 /**
  * @callback onLine
  * @param {string} line The current line.
- * @param {ReturnType<createInterface>} rl The readline interface.
+ * @param {ReturnType<typeof createInterface>} rl The readline interface.
  */
 
 /**
@@ -112,3 +112,94 @@ export const memoize = (fn) => {
     return result;
   };
 };
+
+/**
+ * Waits for a period of time
+ * @param {number} ms The time to wait in milliseconds.
+ */
+export const wait = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
+
+/**
+ * Sends a message to a Discord-compatible webhook.
+ * @param {url|string} url The webhook URL.
+ * @param {string} message The message to be sent.
+ * @returns {Promise}
+ */
+async function sendMessageToWebhook(url, message) {
+  // Create the payload object with the message
+  // The message is provided as 2 different properties to improve compatibility with webhook servers outside Discord
+  const payload = { content: message, body: message };
+
+  // Send a POST request to the webhook url with the payload as JSON
+  try {
+    const response = await fetch(url, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(payload),
+    });
+
+    // Check if the request was successful
+    if (!response.ok) {
+      throw new Error(`HTTP error! Status: ${response.status}`);
+    } else {
+      return true;
+    }
+  } catch (error) {
+    console.error('Error sending message to webhook:', error);
+    return false;
+  }
+}
+
+/**
+ * Sends a CGPS notification to a Discord-compatible webhook.
+ * Automatically checks if the webhook URL exists.
+ * @param {string} msg The message to be sent.
+ * @returns {Promise}
+ */
+export async function notifyWebhook(msg) {
+  // Check if the webhook URL exists
+  const webhook_url = process.env.DISCORD_WEBHOOK_URL;
+
+  if (webhook_url && webhook_url.startsWith('http')) {
+    // Send the message to the webhook
+    try {
+      await sendMessageToWebhook(webhook_url, `CGPS: ${msg}`);
+    } catch (e) {
+      console.error('Error sending message to Discord webhook:', e);
+    }
+  }
+  // Not logging the lack of a webhook URL since it's not a feature everyone would use
+}
+
+/**
+ * Fetches with retry
+ * @param  {Parameters<typeof fetch>} args
+ */
+export const fetchRetry = async (...args) => {
+  let attempts = 0;
+  let maxAttempts = 50;
+  let response;
+
+  while (attempts < maxAttempts) {
+    try {
+      response = await fetch(...args);
+
+      if (!response.ok) {
+        throw new Error(`HTTP error! Status: ${response.status}`);
+      }
+
+      return response;
+    } catch (error) {
+      attempts++;
+      console.warn(`An error occured while making a web request: "${error}", retrying. Attempt ${attempts} of ${maxAttempts}.\nTHIS IS NORMAL IN MOST CIRCUMSTANCES. Refrain from reporting this as a bug unless the script doesn't automatically recover after several attempts.`);
+
+      if (attempts >= maxAttempts) {
+        // Send a message to the Discord webhook if it exists
+        await notifyWebhook(`An HTTP error has occurred (${response ? response.status : "unknown status"}) while making a web request. Please check the logs for further details.`);
+        throw error;
+      }
+
+      await wait(8000);
+    }
+  }
+}
